+++ app/Http/Controllers/MicropubController.php
@@ -109,7 +109,7 @@ class MicropubController extends Controller
             //we have a valid token, is `syndicate-to` set?
             if ($request->input('q') === 'syndicate-to') {
                 $content = http_build_query([
-                    'mp-syndicate-to' => 'twitter.com/jonnybarnes',
+                    'syndicate-to' => 'twitter.com/jonnybarnes',
                 ]);
 
                 return (new Response($content, 200))