+++ app/Jobs/SendWebMentions.php
@@ -2,6 +2,7 @@
 
 namespace App\Jobs;
 
+use App\Note;
 use GuzzleHttp\Client;
 use Illuminate\Queue\SerializesModels;
 use Illuminate\Queue\InteractsWithQueue;
@@ -11,35 +12,41 @@ class SendWebMentions extends Job implements ShouldQueue
 {
     use InteractsWithQueue, SerializesModels;
 
-    protected $url;
-    protected $source;
+    protected $note;
 
     /**
-     * Create a new job instance.
+     * Create the job instance, inject dependencies.
      *
+     * @param  Note $note
      * @return void
      */
-    public function __construct($url, $source)
+    public function __construct(Note $note)
     {
-        $this->url = $url;
-        $this->source = $source;
+        $this->note = $note;
     }
 
     /**
      * Execute the job.
      *
+     * @param  \GuzzleHttp\Client $guzzle
      * @return void
      */
-    public function handle(Client $client)
+    public function handle(Client $guzzle)
     {
-        $endpoint = $this->discoverWebmentionEndpoint($this->url, $client);
-        if ($endpoint) {
-            $client->post($endpoint, [
-                'form_params' => [
-                    'source' => $this->source,
-                    'target' => $this->url,
-                ],
-            ]);
+        //grab the URLs
+        $urlsInReplyTo = explode(' ', $this->note->in_reply_to);
+        $urlsNote = $this->getLinks($this->note->note);
+        $urls = array_filter(array_merge($urlsInReplyTo, $urlsNote)); //filter out none URLs
+        foreach ($urls as $url) {
+            $endpoint = $this->discoverWebmentionEndpoint($url, $guzzle);
+            if ($endpoint) {
+                $guzzle->post($endpoint, [
+                    'form_params' => [
+                        'source' => $this->note->longurl,
+                        'target' => $url,
+                    ],
+                ]);
+            }
         }
     }
 
@@ -47,14 +54,19 @@ class SendWebMentions extends Job implements ShouldQueue
      * Discover if a URL has a webmention endpoint.
      *
      * @param  string  The URL
-     * @param  \GuzzleHttp\Client $client
+     * @param  \GuzzleHttp\Client $guzzle
      * @return string  The webmention endpoint URL
      */
-    private function discoverWebmentionEndpoint($url, $client)
+    private function discoverWebmentionEndpoint($url, $guzzle)
     {
+        //let’s not send webmentions to myself
+        if (parse_url($url, PHP_URL_HOST) == env('LONG_URL', 'localhost')) {
+            return false;
+        }
+
         $endpoint = null;
 
-        $response = $client->get($url);
+        $response = $guzzle->get($url);
         //check HTTP Headers for webmention endpoint
         $links = \GuzzleHttp\Psr7\parse_header($response->getHeader('Link'));
         foreach ($links as $link) {
@@ -82,5 +94,24 @@ class SendWebMentions extends Job implements ShouldQueue
         }
 
         return false;
+    }
+
+    /**
+     * Get the URLs from a note.
+     *
+     * @param  string $html
+     * @return array  $urls
+     */
+    private function getLinks($html)
+    {
+        $urls = [];
+        $dom = new \DOMDocument();
+        $dom->loadHTML($html);
+        $anchors = $dom->getElementsByTagName('a');
+        foreach ($anchors as $anchor) {
+            $urls[] = ($anchor->hasAttribute('href')) ? $anchor->getAttribute('href') : false;
+        }
+
+        return $urls;
     }
 }