+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Version 0.0.5 (2016-06-23)
+  - Automatically send webmentions
+  - Change `mp-syndicate-to` to `syndicate-to`
+
 ## Version 0.0.4 (2016-06-21)
   - Move bower components into their own subdir
   - Move my js into `resources/`, apply an eslint pre-commit hook